python-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Embedding & Reranking
sentence-transformers==3.3.1
torch==2.5.1

# Vector search
faiss-cpu==1.9.0.post1
numpy==1.26.4

# LLM clients
openai==1.58.1
httpx==0.28.1

# Data processing
pandas==2.2.3

# Database
aiosqlite==0.20.0

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
